# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/bookings/{bookingId}/cancellation:
    post:
      operationId: openAdminBookingCancellation
      tags: [excursions]
      summary: Apre un caso di cancellazione amministrativa
      description: >-
        Registra un caso pendente senza annullare ancora la prenotazione. Il
        clientCommandId deve restare stabile durante i retry dello stesso
        comando.
      security:
        - cookieAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCancellationOpenInput"
      responses:
        "200":
          description: Retry idempotente di un comando già registrato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCancellationOpenResponse"
        "201":
          description: Caso amministrativo creato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCancellationOpenResponse"
        "400":
          description: Identificativo comando non valido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Prenotazione non trovata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Prenotazione già annullata o altro caso pendente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/bookings/{bookingId}/cancellation/resolve:
    post:
      operationId: resolveAdminBookingCancellation
      tags: [excursions]
      summary: Approva o rifiuta una cancellazione
      description: Risolve il caso pendente identificato esplicitamente nel body.
      security:
        - cookieAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCancellationResolutionInput"
      responses:
        "200":
          description: Decisione applicata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCancellationResolutionResponse"
        "400":
          description: Decisione o importo non validi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Prenotazione o caso non trovato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Caso non più pendente o conflitto finanziario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refunds/{refundId}/complete-manually:
    post:
      operationId: completeAdminRefundManually
      tags: [excursions]
      summary: Registra il completamento di un rimborso offline
      security:
        - cookieAuth: []
      parameters:
        - name: refundId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminManualRefundCompletionInput"
      responses:
        "200":
          description: Rimborso offline registrato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminManualRefundCompletionResponse"
        "400":
          description: Riferimento mancante o non valido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Rimborso non trovato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Rimborso non completabile manualmente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Leggi impostazioni
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Impostazioni correnti
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Aggiorna impostazioni
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Impostazioni aggiornate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    AdminCancellationOpenInput:
      type: object
      properties:
        clientCommandId:
          type: string
          format: uuid
          description: Identificativo stabile durante i retry dello stesso comando
        reason:
          type: string
          maxLength: 1000
          nullable: true
      required: [clientCommandId]
    AdminCancellationOpenResponse:
      type: object
      properties:
        ok:
          type: boolean
        cancellationCaseId:
          type: string
          format: uuid
        status:
          type: string
        alreadyOpen:
          type: boolean
      required: [ok, cancellationCaseId, status, alreadyOpen]
    AdminCancellationResolutionInput:
      type: object
      properties:
        cancellationCaseId:
          type: string
          format: uuid
        decision:
          type: string
          enum: [approve, reject]
        refundAmountCents:
          type: integer
          minimum: 0
        note:
          type: string
          maxLength: 2000
          nullable: true
          description: Obbligatoria per un rifiuto o quando il rimborso approvato applica una penale
      required: [cancellationCaseId, decision]
    AdminCancellationResolutionResponse:
      type: object
      properties:
        ok:
          type: boolean
        cancellationCaseId:
          type: string
          format: uuid
        status:
          type: string
          enum: [approved, rejected, completed]
        refundIds:
          type: array
          items:
            type: string
            format: uuid
        manualRefundIds:
          type: array
          items:
            type: string
            format: uuid
      required: [ok, cancellationCaseId, status, refundIds, manualRefundIds]
    AdminManualRefundCompletionInput:
      type: object
      properties:
        reference:
          type: string
          minLength: 1
          maxLength: 500
      required: [reference]
    AdminManualRefundCompletionResponse:
      type: object
      properties:
        ok:
          type: boolean
        bookingId:
          type: string
          format: uuid
        cancellationCaseId:
          type: string
          format: uuid
        alreadyCompleted:
          type: boolean
      required: [ok, bookingId, cancellationCaseId, alreadyCompleted]
...
    SettingsResponse:
      type: object
      properties:
        payment_iban:
          type: string
          nullable: true
        payment_beneficiary:
          type: string
          nullable: true
        payment_bank:
          type: string
          nullable: true
        payment_notes:
          type: string
          nullable: true
        deposit_percentage:
          type: string
          nullable: true
        excursion_card_payments_enabled:
          type: string
          nullable: true
        future_card_charge_enabled:
          type: string
          nullable: true
        future_card_charge_consent_version:
          type: string
          nullable: true
        card_checkout_hold_minutes:
          type: string
          nullable: true
        payment_grace_minutes:
          type: string
          nullable: true
        payment_deadline_bank_hours:
          type: string
          nullable: true
        payment_deadline_office_hours:
          type: string
          nullable: true
        payment_deadline_balance_hours:
          type: string
          nullable: true
        payment_deadline_near_departure_hours:
          type: string
          nullable: true
        full_payment_only_days_before:
          type: string
          nullable: true
        auto_release_seats_on_expiry:
          type: string
          nullable: true
        office_address:
          type: string
          nullable: true
        office_opening_hours:
          type: string
          nullable: true
        terms_policy_version:
          type: string
          nullable: true
        privacy_policy_version:
          type: string
          nullable: true
        media_policy_version:
          type: string
          nullable: true
        adult_min_age:
          type: string
          nullable: true
    SettingsInput:
      type: object
      properties:
        payment_iban:
          type: string
        payment_beneficiary:
          type: string
        payment_bank:
          type: string
        payment_notes:
          type: string
        deposit_percentage:
          type: string
        excursion_card_payments_enabled:
          type: string
        future_card_charge_enabled:
          type: string
        future_card_charge_consent_version:
          type: string
        card_checkout_hold_minutes:
          type: string
        payment_grace_minutes:
          type: string
        payment_deadline_bank_hours:
          type: string
        payment_deadline_office_hours:
          type: string
        payment_deadline_balance_hours:
          type: string
        payment_deadline_near_departure_hours:
          type: string
        full_payment_only_days_before:
          type: string
        auto_release_seats_on_expiry:
          type: string
        office_address:
          type: string
        office_opening_hours:
          type: string
        terms_policy_version:
          type: string
        privacy_policy_version:
          type: string
        media_policy_version:
          type: string
        adult_min_age:
          type: string
...